+++ src/com/donny/dendrofinance/account/Account.java
@@ -23,6 +23,155 @@ public class Account implements ExportableToJson, Serializable {
             selfIncName,
             fixedAssetsTypeName, receiveTypeName,
             appreciationName, depreciationName;
+    public static boolean specialAltered = false;
+
+    public static void changeSpecial(String current, String replacement) {
+        if (portfolioName.equals(current)) {
+            portfolioName = replacement;
+            specialAltered = true;
+        }
+        if (stockName.equals(current)) {
+            stockName = replacement;
+            specialAltered = true;
+        }
+        if (inventoryName.equals(current)) {
+            inventoryName = replacement;
+            specialAltered = true;
+        }
+        if (fiatName.equals(current)) {
+            fiatName = replacement;
+            specialAltered = true;
+        }
+        if (cryptoName.equals(current)) {
+            cryptoName = replacement;
+            specialAltered = true;
+        }
+        if (tradIncName.equals(current)) {
+            tradIncName = replacement;
+            specialAltered = true;
+        }
+        if (tradExpName.equals(current)) {
+            tradExpName = replacement;
+            specialAltered = true;
+        }
+        if (cgStockName.equals(current)) {
+            cgStockName = replacement;
+            specialAltered = true;
+        }
+        if (cgInventoryName.equals(current)) {
+            cgInventoryName = replacement;
+            specialAltered = true;
+        }
+        if (cgFiatName.equals(current)) {
+            cgFiatName = replacement;
+            specialAltered = true;
+        }
+        if (cgCryptoName.equals(current)) {
+            cgCryptoName = replacement;
+            specialAltered = true;
+        }
+        if (clStockName.equals(current)) {
+            clStockName = replacement;
+            specialAltered = true;
+        }
+        if (clInventoryName.equals(current)) {
+            clInventoryName = replacement;
+            specialAltered = true;
+        }
+        if (clFiatName.equals(current)) {
+            clFiatName = replacement;
+            specialAltered = true;
+        }
+        if (clCryptoName.equals(current)) {
+            clCryptoName = replacement;
+            specialAltered = true;
+        }
+        if (cgltStockName.equals(current)) {
+            cgltStockName = replacement;
+            specialAltered = true;
+        }
+        if (cgltInventoryName.equals(current)) {
+            cgltInventoryName = replacement;
+            specialAltered = true;
+        }
+        if (cgltFiatName.equals(current)) {
+            cgltFiatName = replacement;
+            specialAltered = true;
+        }
+        if (cgltCryptoName.equals(current)) {
+            cgltCryptoName = replacement;
+            specialAltered = true;
+        }
+        if (clltStockName.equals(current)) {
+            clltStockName = replacement;
+            specialAltered = true;
+        }
+        if (clltInventoryName.equals(current)) {
+            clltInventoryName = replacement;
+            specialAltered = true;
+        }
+        if (clltFiatName.equals(current)) {
+            clltFiatName = replacement;
+            specialAltered = true;
+        }
+        if (clltCryptoName.equals(current)) {
+            clltCryptoName = replacement;
+            specialAltered = true;
+        }
+        if (selfIncName.equals(current)) {
+            selfIncName = replacement;
+            specialAltered = true;
+        }
+        if (fixedAssetsTypeName.equals(current)) {
+            fixedAssetsTypeName = replacement;
+            specialAltered = true;
+        }
+        if (receiveTypeName.equals(current)) {
+            receiveTypeName = replacement;
+            specialAltered = true;
+        }
+        if (appreciationName.equals(current)) {
+            appreciationName = replacement;
+            specialAltered = true;
+        }
+        if (depreciationName.equals(current)) {
+            depreciationName = replacement;
+            specialAltered = true;
+        }
+    }
+
+    public static JsonObject specialExport() throws JsonFormattingException {
+        JsonObject object = new JsonObject();
+        object.put("portfolio", new JsonString(portfolioName));
+        object.put("stock", new JsonString(stockName));
+        object.put("inventory", new JsonString(inventoryName));
+        object.put("fiat", new JsonString(fiatName));
+        object.put("crypto", new JsonString(cryptoName));
+        object.put("trading-income", new JsonString(tradIncName));
+        object.put("trading-expense", new JsonString(tradExpName));
+        object.put("stock-gain", new JsonString(cgStockName));
+        object.put("inventory-gain", new JsonString(cgInventoryName));
+        object.put("fiat-gain", new JsonString(cgFiatName));
+        object.put("crypto-gain", new JsonString(cgCryptoName));
+        object.put("stock-loss", new JsonString(clStockName));
+        object.put("inventory-loss", new JsonString(clInventoryName));
+        object.put("fiat-loss", new JsonString(clFiatName));
+        object.put("crypto-loss", new JsonString(clCryptoName));
+        object.put("stock-gain-long-run", new JsonString(cgltStockName));
+        object.put("inventory-gain-long-run", new JsonString(cgltInventoryName));
+        object.put("fiat-gain-long-run", new JsonString(cgltFiatName));
+        object.put("crypto-gain-long-run", new JsonString(cgltCryptoName));
+        object.put("stock-loss-long-run", new JsonString(clltStockName));
+        object.put("inventory-loss-long-run", new JsonString(clltInventoryName));
+        object.put("fiat-loss-long-run", new JsonString(clltFiatName));
+        object.put("crypto-loss-long-run", new JsonString(clltCryptoName));
+        object.put("self-employment-income", new JsonString(selfIncName));
+        object.put("fixed-assets-type", new JsonString(fixedAssetsTypeName));
+        object.put("receivables-type", new JsonString(receiveTypeName));
+        object.put("appreciation", new JsonString(appreciationName));
+        object.put("depreciation", new JsonString(depreciationName));
+        return object;
+    }
 
     public final boolean EXPORT;
     public final Exchange EXCHANGE;
@@ -122,6 +271,38 @@ public class Account implements ExportableToJson, Serializable {
                 }
             }
         }
+        return false;
+    }
+
+    public boolean inSpecial() {
+        if (NAME.equalsIgnoreCase(portfolioName)) return true;
+        if (NAME.equalsIgnoreCase(stockName)) return true;
+        if (NAME.equalsIgnoreCase(inventoryName)) return true;
+        if (NAME.equalsIgnoreCase(fiatName)) return true;
+        if (NAME.equalsIgnoreCase(cryptoName)) return true;
+        if (NAME.equalsIgnoreCase(tradIncName)) return true;
+        if (NAME.equalsIgnoreCase(tradExpName)) return true;
+        if (NAME.equalsIgnoreCase(cgStockName)) return true;
+        if (NAME.equalsIgnoreCase(cgInventoryName)) return true;
+        if (NAME.equalsIgnoreCase(cgFiatName)) return true;
+        if (NAME.equalsIgnoreCase(cgCryptoName)) return true;
+        if (NAME.equalsIgnoreCase(clStockName)) return true;
+        if (NAME.equalsIgnoreCase(clInventoryName)) return true;
+        if (NAME.equalsIgnoreCase(clFiatName)) return true;
+        if (NAME.equalsIgnoreCase(clCryptoName)) return true;
+        if (NAME.equalsIgnoreCase(cgltStockName)) return true;
+        if (NAME.equalsIgnoreCase(cgltInventoryName)) return true;
+        if (NAME.equalsIgnoreCase(cgltFiatName)) return true;
+        if (NAME.equalsIgnoreCase(cgltCryptoName)) return true;
+        if (NAME.equalsIgnoreCase(clltStockName)) return true;
+        if (NAME.equalsIgnoreCase(clltInventoryName)) return true;
+        if (NAME.equalsIgnoreCase(clltFiatName)) return true;
+        if (NAME.equalsIgnoreCase(clltCryptoName)) return true;
+        if (NAME.equalsIgnoreCase(selfIncName)) return true;
+        if (NAME.equalsIgnoreCase(fixedAssetsTypeName)) return true;
+        if (NAME.equalsIgnoreCase(receiveTypeName)) return true;
+        if (NAME.equalsIgnoreCase(appreciationName)) return true;
+        if (NAME.equalsIgnoreCase(depreciationName)) return true;
         return false;
     }